requirements.txt
# Core Deep Learning
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1

# YOLO Framework
ultralytics==8.1.25

# Data Processing & Image Processing
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.21.0
scipy==1.11.4

# ML & Utilities
scikit-learn==1.3.2
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1

# Web Framework (Flask) - Optional
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Configuration & Logging
python-dotenv==1.0.0
pyyaml==6.0.1

# Metrics & Evaluation
torchmetrics==1.3.0

# Development & Debugging
jupyter==1.0.0
ipython==8.18.1
black==23.12.0
flake8==6.1.0
